Update docs and reorganizes modules
