+++ docs/src/main/tut/metrics-reporting.md
@@ -6,11 +6,16 @@ permalink: /metrics-reporting
 
 # Metrics Reporting
 
-Currently, [mu] provides two different ways to monitor [gRPC] services: `Prometheus` and `Dropwizard` (using the `Prometheus` extension). The usage is quite similar for both.
+Currently, [mu] provides two different ways to monitor [gRPC] services: `Prometheus` and `Dropwizard`. The usage is quite similar for both.
 
 ## Monitor Server Calls
 
-In order to monitor the RPC calls on the server side, it's necessary to intercept them. We'll see how to do this in the next code fragment:
+In order to monitor the RPC calls on the server side we need two things:
+
+* A `MetricsOps` implementation. `MetricsOps` is an algebra located in the `internal-core` module with the needed operations for registering metrics. [mu] provides two implementations, one for `Prometheus` and another one for `Dropwizard` but you can provide your own.
+* A `MetricsServerInterceptor`. [mu] provides an interceptor that receives a `MetricsOps` as an argument and register server metrics.
+
+Let's see how to register server metrics using `Prometheus` in the following fragment.
 
 ```tut:invisible
 trait CommonRuntime {
@@ -57,31 +62,25 @@ class ServiceHandler[F[_]: Applicative] extends Greeter[F] {
 
 ```tut:silent
 import cats.effect.IO
+import higherkindness.mu.rpc.prometheus.PrometheusMetrics
 import higherkindness.mu.rpc.server._
-import higherkindness.mu.rpc.prometheus.shared.Configuration
-import higherkindness.mu.rpc.prometheus.server.MonitoringServerInterceptor
+import higherkindness.mu.rpc.server.interceptors.implicits._
+import higherkindness.mu.rpc.server.metrics.MetricsServerInterceptor
 import io.prometheus.client.CollectorRegistry
 import service._
 
 object InterceptingServerCalls extends CommonRuntime {
 
-  import higherkindness.mu.rpc.interceptors.implicits._
-
   lazy val cr: CollectorRegistry = new CollectorRegistry()
-  lazy val monitorInterceptor = MonitoringServerInterceptor(
-    Configuration.defaultBasicMetrics.withCollectorRegistry(cr)
-  )
 
   implicit val greeterServiceHandler: ServiceHandler[IO] = new ServiceHandler[IO]
-
-  // The Greeter service is the service defined in the Core concepts section
-  val grpcConfigs: IO[List[GrpcConfig]] = 
-    Greeter.bindService[IO]
-      .map(_.interceptWith(monitorInterceptor))
-      .map(AddService)
-      .map(List(_))
-
-  val server: IO[GrpcServer[IO]] = grpcConfigs.flatMap(GrpcServer.default[IO](8080, _))
+  
+  val server: IO[GrpcServer[IO]] = for {
+    metricsOps <- PrometheusMetrics[IO](cr, "server")
+    service    <- Greeter.bindService[IO]
+    grpcConfig = AddService(service.interceptWith(MetricsServerInterceptor(metricsOps)))
+    server     <- GrpcServer.default[IO](8080, List(grpcConfig))
+  } yield server
 
 }
 ```
@@ -95,29 +94,23 @@ import cats.effect.{IO, Resource}
 import higherkindness.mu.rpc._
 import higherkindness.mu.rpc.config._
 import higherkindness.mu.rpc.channel._
+import higherkindness.mu.rpc.channel.metrics.MetricsChannelInterceptor
 import higherkindness.mu.rpc.config.channel._
+import io.prometheus.client.CollectorRegistry
 import service._
-
-import higherkindness.mu.rpc.prometheus.shared.Configuration
-import higherkindness.mu.rpc.prometheus.client.MonitoringClientInterceptor
 
 object InterceptingClientCalls extends CommonRuntime {
 
-  val channelFor: ChannelFor =
-    ConfigForAddress[IO]("rpc.host", "rpc.port").unsafeRunSync
+  lazy val cr: CollectorRegistry = new CollectorRegistry()
 
-  implicit val serviceClient: Resource[IO, Greeter[IO]] =
-    Greeter.client[IO](
-      channelFor = channelFor,
-      channelConfigList = List(
-        UsePlaintext(),
-        AddInterceptor(
-          MonitoringClientInterceptor(
-            Configuration.defaultBasicMetrics
-          )
-        )
-      )
-    )
+  implicit val serviceClient: Resource[IO, Greeter[IO]] = 
+    for {
+      channelFor    <- Resource.liftF(ConfigForAddress[IO]("rpc.host", "rpc.port"))
+      metricsOps    <- Resource.liftF(PrometheusMetrics[IO](cr, "client"))
+      serviceClient <- Greeter.client[IO](
+        channelFor = channelFor, 
+        channelConfigList = List(UsePlaintext(), AddInterceptor(MetricsChannelInterceptor(metricsOps))))
+    } yield serviceClient
 }
 ```
 
@@ -125,15 +118,14 @@ That is how we use `Prometheus` to monitor both [gRPC] ends.
 
 ## Dropwizard Metrics
 
-The usage the same as before, but in this case we need to put an instance of `com.codahale.metrics.MetricRegistry` in our code. Then using the _Dropwizard_ integration that _Prometheus_ already provides (`DropwizardExports`) we can associate it with the collector registry:
+The usage the same as before, but in this case we need to create a `Dropwizard` backed `MetricsOps` 
 
 ```tut:silent
+import cats.effect.IO
 import com.codahale.metrics.MetricRegistry
-import io.prometheus.client.dropwizard.DropwizardExports
+import higherkindness.mu.rpc.dropwizard.DropWizardMetrics
 
-val metrics: MetricRegistry      = new MetricRegistry
-val configuration: Configuration = Configuration.defaultBasicMetrics
-configuration.collectorRegistry.register(new DropwizardExports(metrics))
+val metricsOps = DropWizardMetrics[IO](new MetricRegistry)
 ```
 
 [RPC]: https://en.wikipedia.org/wiki/Remote_procedure_call
